requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.0.1
altair == 4.2.0
dotenv == 0.0.5
numpy == 1.21.5
pandas == 1.4.2
streamlit == 1.15.2
